Cleanup enums [D]
