+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridParentRowsLabel.cs
@@ -2,43 +2,33 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Windows.Forms {
-        /// <include file='doc\DataGridParentRowsLabel.uex' path='docs/doc[@for="DataGridParentRowsLabelStyle"]/*' />
+namespace System.Windows.Forms
+{
+    /// <devdoc>
+    /// Specifies how parent row labels of a DataGrid
+    /// control are displayed.
+    /// </devdoc>
+    public enum DataGridParentRowsLabelStyle
+    {
         /// <devdoc>
-        ///    <para>
-        ///       Specifies how parent row labels of a DataGrid
-        ///       control are displayed.
-        ///    </para>
+        /// Display no parent row labels.
         /// </devdoc>
-        public enum DataGridParentRowsLabelStyle {
-            /// <include file='doc\DataGridParentRowsLabel.uex' path='docs/doc[@for="DataGridParentRowsLabelStyle.None"]/*' />
-            /// <devdoc>
-            ///    <para>
-            ///       Display no parent row labels.
-            ///    </para>
-            /// </devdoc>
-            None              = 0,
-            /// <include file='doc\DataGridParentRowsLabel.uex' path='docs/doc[@for="DataGridParentRowsLabelStyle.TableName"]/*' />
-            /// <devdoc>
-            ///    <para>
-            ///       Displaya the parent table name.
-            ///    </para>
-            /// </devdoc>
-            TableName            = 1,
-            /// <include file='doc\DataGridParentRowsLabel.uex' path='docs/doc[@for="DataGridParentRowsLabelStyle.ColumnName"]/*' />
-            /// <devdoc>
-            ///    <para>
-            ///       Displaya the parent column name.
-            ///    </para>
-            /// </devdoc>
-            ColumnName           = 2,
-            /// <include file='doc\DataGridParentRowsLabel.uex' path='docs/doc[@for="DataGridParentRowsLabelStyle.Both"]/*' />
-            /// <devdoc>
-            ///    <para>
-            ///       Displays
-            ///       both the parent table and column names.
-            ///    </para>
-            /// </devdoc>
-            Both  = 3,
-        }
+        None = 0,
+
+        /// <devdoc>
+        /// Displaya the parent table name.
+        /// </devdoc>
+        TableName = 1,
+
+        /// <devdoc>
+        /// Displaya the parent column name.
+        /// </devdoc>
+        ColumnName = 2,
+
+        /// <devdoc>
+        /// Displays
+        /// both the parent table and column names.
+        /// </devdoc>
+        Both = 3,
+    }
 }
