+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewAdvancedCellBorderStyle.cs
@@ -4,34 +4,15 @@
 
 namespace System.Windows.Forms
 {
-    using System;
-    using System.ComponentModel;
-
-    /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.DataGridViewAdvancedCellBorderStyle"]/*' />
     public enum DataGridViewAdvancedCellBorderStyle
     {
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.NotSet"]/*' />
         NotSet = 0,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.None"]/*' />
         None = 1,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.Single"]/*' />
         Single = 2,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.Inset"]/*' />
         Inset = 3,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.InsetDouble"]/*' />
         InsetDouble = 4,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.Outset"]/*' />
         Outset = 5,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.OutsetDouble"]/*' />
         OutsetDouble = 6,
-
-        /// <include file='doc\DataGridViewAdvancedCellBorderStyle.uex' path='docs/doc[@for="DataGridViewAdvancedCellBorderStyle.OutsetPartial"]/*' />
         OutsetPartial = 7
     }
 }
