+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewCellBorderStyle.cs
@@ -4,40 +4,18 @@
 
 namespace System.Windows.Forms
 {
-    /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.DataGridViewCellBorderStyle"]/*' />
     public enum DataGridViewCellBorderStyle
     {
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.Custom"]/*' />
         Custom = 0,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.Single"]/*' />
         Single = 1,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.Raised"]/*' />
         Raised = 2,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.Sunken"]/*' />
         Sunken = 3,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.None"]/*' />
         None = 4,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.SingleVertical"]/*' />
         SingleVertical = 5,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.RaisedVertical"]/*' />
         RaisedVertical = 6,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.SunkenVertical"]/*' />
         SunkenVertical = 7,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.SingleHorizontal"]/*' />
         SingleHorizontal = 8,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.RaisedHorizontal"]/*' />
         RaisedHorizontal = 9,
-
-        /// <include file='doc\DataGridViewCellBorderStyle.uex' path='docs/doc[@for="DataGridViewCellBorderStyle.SunkenHorizontal"]/*' />
         SunkenHorizontal = 10
     }
 }
