+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewColumnHeadersHeightSizeMode.cs
@@ -2,13 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Diagnostics.CodeAnalysis;
+
 namespace System.Windows.Forms
 {
-    public enum DataGridViewClipboardCopyMode
+    public enum DataGridViewColumnHeadersHeightSizeMode
     {
-        Disable = 0,
-        EnableWithAutoHeaderText,
-        EnableWithoutHeaderText,
-        EnableAlwaysIncludeHeaderText
+        EnableResizing = 0,
+        DisableResizing,
+        AutoSize
     }
 }
