+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewColumnSortMode.cs
@@ -4,17 +4,10 @@
 
 namespace System.Windows.Forms
 {
-    using System;
-    using System.ComponentModel;
-
-    /// <include file='doc\DataGridViewColumnSortMode.uex' path='docs/doc[@for="DataGridViewColumnSortMode"]/*' />
     public enum DataGridViewColumnSortMode
     {
-        /// <include file='doc\DataGridViewColumnSortMode.uex' path='docs/doc[@for="DataGridViewColumnSortMode.NotSortable"]/*' />
         NotSortable,
-        /// <include file='doc\DataGridViewColumnSortMode.uex' path='docs/doc[@for="DataGridViewColumnSortMode.Automatic"]/*' />
         Automatic,
-        /// <include file='doc\DataGridViewColumnSortMode.uex' path='docs/doc[@for="DataGridViewColumnSortMode.Programmatic"]/*' />
         Programmatic
     }
 }
