+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewComboBoxDisplayStyle.cs
@@ -4,16 +4,10 @@
 
 namespace System.Windows.Forms
 {
-    /// <include file='doc\DataGridViewComboBoxDisplayStyle.uex' path='docs/doc[@for="DataGridViewComboBoxDisplayStyle.DataGridViewComboBoxDisplayStyle"]/*' />
     public enum DataGridViewComboBoxDisplayStyle
     {
-        /// <include file='doc\DataGridViewComboBoxDisplayStyle.uex' path='docs/doc[@for="DataGridViewComboBoxDisplayStyle.ComboBox"]/*' />
         ComboBox,
-
-        /// <include file='doc\DataGridViewComboBoxDisplayStyle.uex' path='docs/doc[@for="DataGridViewComboBoxDisplayStyle.DropDownButton"]/*' />
         DropDownButton,
-
-        /// <include file='doc\DataGridViewComboBoxDisplayStyle.uex' path='docs/doc[@for="DataGridViewComboBoxDisplayStyle.Nothing"]/*' />
         Nothing
     }
 }
