+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewHeaderBorderStyle.cs
@@ -4,22 +4,12 @@
 
 namespace System.Windows.Forms
 {
-    /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.DataGridViewHeaderBorderStyle"]/*' />
     public enum DataGridViewHeaderBorderStyle
     {
-        /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.Custom"]/*' />
         Custom = 0,
-
-        /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.Single"]/*' />
         Single = 1,
-
-        /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.Raised"]/*' />
         Raised = 2,
-
-        /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.Sunken"]/*' />
         Sunken = 3,
-
-        /// <include file='doc\DataGridViewHeaderBorderStyle.uex' path='docs/doc[@for="DataGridViewHeaderBorderStyle.None"]/*' />
         None = 4
     }
 }
