+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewHitTestType.cs
@@ -4,34 +4,18 @@
 
 namespace System.Windows.Forms
 {
-    using System;
-    using System.ComponentModel;
-
-    /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType"]/*' />
     /// <devdoc>
-    /// <para>Specifies the part of the <see cref='System.Windows.Forms.DataGridView'/> control where the mouse is.</para>
+    /// Specifies the part of the <see cref='System.Windows.Forms.DataGridView'/>
+    /// control where the mouse is.
     /// </devdoc>
     public enum DataGridViewHitTestType
     {
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.None"]/*' />
-        None          = 0,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.Cell"]/*' />
-        Cell          = 1,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.ColumnHeader"]/*' />
-        ColumnHeader  = 2,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.RowHeader"]/*' />
-        RowHeader     = 3,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.TopLeftHeader"]/*' />
+        None = 0,
+        Cell = 1,
+        ColumnHeader = 2,
+        RowHeader = 3,
         TopLeftHeader = 4,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.HorizontalScrollBar"]/*' />
         HorizontalScrollBar = 5,
-
-        /// <include file='doc\DataGridViewHitTestType.uex' path='docs/doc[@for="DataGridViewHitTestType.VerticalScrollBar"]/*' />
         VerticalScrollBar = 6
     }
 }
