+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewImageCellLayout.cs
@@ -4,19 +4,14 @@
 
 namespace System.Windows.Forms
 {
-    /// <include file='doc\DataGridViewImageCellLayout.uex' path='docs/doc[@for="DataGridViewImageCellLayout.DataGridViewImageCellLayout"]/*' />
     public enum DataGridViewImageCellLayout
     {
-        /// <include file='doc\DataGridViewImageCellLayout.uex' path='docs/doc[@for="DataGridViewImageCellLayout.NotSet"]/*' />
         NotSet = 0,
 
-        /// <include file='doc\DataGridViewImageCellLayout.uex' path='docs/doc[@for="DataGridViewImageCellLayout.Normal"]/*' />
         Normal,
 
-        /// <include file='doc\DataGridViewImageCellLayout.uex' path='docs/doc[@for="DataGridViewImageCellLayout.Stretch"]/*' />
         Stretch,
 
-        /// <include file='doc\DataGridViewImageCellLayout.uex' path='docs/doc[@for="DataGridViewImageCellLayout.Zoom"]/*' />
         Zoom
     }
 }
