+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewRowHeadersWidthSizeMode.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    public enum DataGridViewRowHeadersWidthSizeMode
+    {
+        EnableResizing = 0,
+        DisableResizing,
+        AutoSizeToAllHeaders,
+        AutoSizeToDisplayedHeaders,
+        AutoSizeToFirstHeader
+    }
+}
