+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewSelectionMode.cs
@@ -4,28 +4,12 @@
 
 namespace System.Windows.Forms
 {
-    using System;
-    using System.ComponentModel;
-
-    /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.DataGridViewSelectionMode"]/*' />
-    /// <devdoc>
-    /// <para></para>
-    /// </devdoc>
     public enum DataGridViewSelectionMode 
     {
-        /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.CellSelect"]/*' />
         CellSelect = 0,
-
-        /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.FullRowSelect"]/*' />
         FullRowSelect = 1,
-
-        /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.FullColumnSelect"]/*' />
         FullColumnSelect = 2,
-
-        /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.RowHeaderSelect"]/*' />
         RowHeaderSelect = 3,
-
-        /// <include file='doc\DataGridViewSelectionMode.uex' path='docs/doc[@for="DataGridViewSelectionMode.ColumnHeaderSelect"]/*' />
         ColumnHeaderSelect = 4
     }
 }
