+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewTriState.cs
@@ -4,19 +4,10 @@
 
 namespace System.Windows.Forms
 {
-    using System;
-    using System.ComponentModel;
-
-    /// <include file='doc\DataGridViewTriState.uex' path='docs/doc[@for="DataGridViewTriState.DataGridViewTriState"]/*' />
     public enum DataGridViewTriState
     {
-        /// <include file='doc\DataGridViewTriState.uex' path='docs/doc[@for="DataGridViewTriState.NotSet"]/*' />
         NotSet = 0,
-
-        /// <include file='doc\DataGridViewTriState.uex' path='docs/doc[@for="DataGridViewTriState.True"]/*' />
         True,
-    
-        /// <include file='doc\DataGridViewTriState.uex' path='docs/doc[@for="DataGridViewTriState.False"]/*' />
         False
     }
 }
