+++ src/System.Windows.Forms/src/System/Windows/Forms/Day.cs
@@ -2,78 +2,50 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Windows.Forms {
-
-    using System.Diagnostics;
-
-    using System;
-    using System.Drawing;
-    using System.ComponentModel;
-    using Microsoft.Win32;
-
-
-    /// <include file='doc\Day.uex' path='docs/doc[@for="Day"]/*' />
+namespace System.Windows.Forms
+{
     /// <devdoc>
-    ///    <para>
-    ///       Specifies the day of the week.
-    ///
-    ///    </para>
+    /// Specifies the day of the week.
     /// </devdoc>
-    public enum Day {
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Monday"]/*' />
+    public enum Day
+    {
         /// <devdoc>
-        ///    <para>
-        ///       The day Monday.
-        ///    </para>
+        /// The day Monday.
         /// </devdoc>
         Monday = 0,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Tuesday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Tuesday.
-        ///    </para>
+        /// The day Tuesday.
         /// </devdoc>
         Tuesday = 1,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Wednesday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Wednesday.
-        ///    </para>
+        /// The day Wednesday.
         /// </devdoc>
         Wednesday = 2,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Thursday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Thursday.
-        ///    </para>
+        /// The day Thursday.
         /// </devdoc>
         Thursday = 3,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Friday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Friday.
-        ///    </para>
+        /// The day Friday.
         /// </devdoc>
         Friday = 4,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Saturday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Saturday.
-        ///    </para>
+        /// The day Saturday.
         /// </devdoc>
         Saturday  = 5,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Sunday"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       The day Sunday.
-        ///    </para>
+        /// The day Sunday.
         /// </devdoc>
         Sunday = 6,
-        /// <include file='doc\Day.uex' path='docs/doc[@for="Day.Default"]/*' />
+
         /// <devdoc>
-        ///    <para>
-        ///       A default day of the week specified by the application.
-        ///    </para>
+        /// A default day of the week specified by the application.
         /// </devdoc>
         Default = 7,
     }
