+++ src/System.Windows.Forms/src/System/Windows/Forms/DialogResult.cs
@@ -2,110 +2,61 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace System.Windows.Forms {
+using System.Runtime.InteropServices;
 
-    using System.Diagnostics;
-
-    using System;
-    using System.Drawing;
-    using System.ComponentModel;
-    using Microsoft.Win32;
-
-
-
-    /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult"]/*' />
+namespace System.Windows.Forms
+{
     /// <devdoc>
-    ///    <para>
-    ///       Specifies identifiers to
-    ///       indicate the return value of a dialog box.
-    ///    </para>
+    /// Specifies identifiers to indicate the return value of a dialog box.
     /// </devdoc>
-    [System.Runtime.InteropServices.ComVisible(true)]
-    public enum DialogResult{
-
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.None"]/*' />
+    [ComVisible(true)]
+    public enum DialogResult
+    {
         /// <devdoc>
-        ///    <para>
-        ///       
-        ///       Nothing is returned from the dialog box. This
-        ///       means that the modal dialog continues running.
-        ///       
-        ///    </para>
+        /// Nothing is returned from the dialog box. This means that the modal
+        /// dialog continues running.
         /// </devdoc>
         None = 0,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.OK"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The
-        ///       dialog box return value is
-        ///       OK (usually sent from a button labeled OK).
-        ///       
-        ///    </para>
+        /// The dialog box return value is OK (usually sent from a button labeled OK).
         /// </devdoc>
         OK = 1,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.Cancel"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The
-        ///       dialog box return value is Cancel (usually sent
-        ///       from a button labeled Cancel).
-        ///       
-        ///    </para>
+        /// The dialog box return value is Cancel (usually sent from a button
+        /// labeled Cancel).
         /// </devdoc>
         Cancel = 2,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.Abort"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The dialog box return value is
-        ///       Abort (usually sent from a button labeled Abort).
-        ///       
-        ///    </para>
+        /// The dialog box return value is Abort (usually sent from a button
+        /// labeled Abort).
         /// </devdoc>
         Abort = 3,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.Retry"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The dialog box return value is
-        ///       Retry (usually sent from a button labeled Retry).
-        ///       
-        ///    </para>
+        /// The dialog box return value is Retry (usually sent from a button
+        /// labeled Retry).
         /// </devdoc>
         Retry = 4,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.Ignore"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The
-        ///       dialog box return value is Ignore (usually sent
-        ///       from a button labeled Ignore).
-        ///       
-        ///    </para>
+        /// The dialog box return value is Ignore (usually sent from a button
+        /// labeled Ignore).
         /// </devdoc>
         Ignore = 5,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.Yes"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The dialog box return value is
-        ///       Yes (usually sent from a button labeled Yes).
-        ///       
-        ///    </para>
+        /// The dialog box return value is Yes (usually sent from a button
+        /// labeled Yes).
         /// </devdoc>
         Yes = 6,
 
-        /// <include file='doc\DialogResult.uex' path='docs/doc[@for="DialogResult.No"]/*' />
         /// <devdoc>
-        ///    <para>
-        ///       The dialog box return value is
-        ///       No (usually sent from a button labeled No).
-        ///       
-        ///    </para>
+        /// The dialog box return value is No (usually sent from a button
+        /// labeled No).
         /// </devdoc>
         No = 7,
-
     }
 }
